Refactor API route mapping into endpoint modules
